web-api/requirements.txt
SQLAlchemy~=2.0.38
pyodbc~=5.2.0
python-dotenv~=1.0.1
Flask~=3.1.0
Flask-Cors~=5.0.0
Faker~=35.2.0
schedule~=1.2.2